refactor: 記事のヘッダーとフッターを別コンポーネントに分割
