+++ web/src/components/ArticleFooter.vue
@@ -1,0 +1,258 @@
+<script lang="ts" setup>
+// Components
+import {
+  SocialShare,
+  SocialShareButton,
+  SocialShareInput,
+  SocialSharePanel,
+  SocialShareSubmit,
+} from '@hiratake/social-share'
+// Icons
+import { ChevronLeftIcon, ChevronRightIcon } from '@heroicons/vue/20/solid'
+import { siFacebook, siMastodon, siMisskey, siTwitter } from 'simple-icons'
+
+const { next, page, prev } = useContent()
+const config = useRuntimeConfig()
+const route = useRoute()
+</script>
+
+<template>
+  <footer v-if="page" class="grid grid-cols-1 gap-16">
+    <div class="grid gap-2">
+      <div class="text-center text-xs">SNSでこのページをシェアする</div>
+      <div class="flex items-center justify-center gap-1">
+        <SocialShare
+          :url="`${config.public.siteUrl}${
+            route.path.endsWith('/') ? route.path : `${route.path}/`
+          }`"
+          :text="`${page?.title || ''} - ${config.public.siteName}`"
+          service="twitter"
+          class="relative"
+        >
+          <SocialShareButton
+            :class="[
+              'flex h-10 w-10 items-center justify-center p-2',
+              'transition-opacity hover:opacity-70',
+            ]"
+            title="Twitterでシェア"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 24 24"
+              class="h-full w-full fill-current text-current"
+            >
+              <path :d="siTwitter.path" />
+            </svg>
+          </SocialShareButton>
+        </SocialShare>
+
+        <SocialShare
+          :url="`${config.public.siteUrl}${
+            route.path.endsWith('/') ? route.path : `${route.path}/`
+          }`"
+          :text="`${page?.title || ''} - ${config.public.siteName}`"
+          service="facebook"
+          class="relative"
+        >
+          <SocialShareButton
+            :class="[
+              'flex h-10 w-10 items-center justify-center p-2',
+              'transition-opacity hover:opacity-70',
+            ]"
+            title="Facebookでシェア"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 24 24"
+              class="h-full w-full fill-current text-current"
+            >
+              <path :d="siFacebook.path" />
+            </svg>
+          </SocialShareButton>
+        </SocialShare>
+
+        <SocialShare
+          :url="`${config.public.siteUrl}${
+            route.path.endsWith('/') ? route.path : `${route.path}/`
+          }`"
+          :text="`${page?.title || ''} - ${config.public.siteName}`"
+          service="mastodon"
+          class="relative"
+        >
+          <SocialShareButton
+            :class="[
+              'flex h-10 w-10 items-center justify-center p-2',
+              'transition-opacity hover:opacity-70',
+            ]"
+            title="Mastodonでシェア"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 24 24"
+              class="h-full w-full fill-current text-current"
+            >
+              <path :d="siMastodon.path" />
+            </svg>
+          </SocialShareButton>
+
+          <Transition
+            enter-active-class="transition duration-200 ease-out"
+            enter-from-class="translate-y-1 opacity-0"
+            enter-to-class="translate-y-0 opacity-100"
+            leave-active-class="transition duration-150 ease-in"
+            leave-from-class="translate-y-0 opacity-100"
+            leave-to-class="translate-y-1 opacity-0"
+          >
+            <SocialSharePanel
+              :class="[
+                'absolute bottom-12 left-1/2 -translate-x-1/2',
+                'grid w-52 gap-1 rounded-lg p-3',
+                'shadow-lg dark:shadow-slate-950/70',
+                'bg-white dark:bg-slate-800',
+              ]"
+            >
+              <label class="select-none text-xs" for="share-mastodon">
+                サーバのドメイン：
+              </label>
+              <div class="flex items-center gap-2">
+                <SocialShareInput
+                  id="share-mastodon"
+                  placeholder="example.com"
+                  :class="[
+                    'shrink grow basis-full',
+                    'w-full rounded px-2 py-1 text-sm outline-none',
+                    'bg-slate-100 dark:bg-slate-700',
+                  ]"
+                />
+                <SocialShareSubmit
+                  :class="[
+                    'shrink-0 rounded border border-slate-300 px-2 py-1 text-sm',
+                  ]"
+                >
+                  投稿
+                </SocialShareSubmit>
+              </div>
+            </SocialSharePanel>
+          </Transition>
+        </SocialShare>
+
+        <SocialShare
+          :url="`${config.public.siteUrl}${
+            route.path.endsWith('/') ? route.path : `${route.path}/`
+          }`"
+          :text="`${page?.title || ''} - ${config.public.siteName}`"
+          service="misskey"
+          class="relative"
+        >
+          <SocialShareButton
+            :class="[
+              'flex h-10 w-10 items-center justify-center p-2',
+              'transition-opacity hover:opacity-70',
+            ]"
+            title="Misskeyでシェア"
+          >
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              viewBox="0 0 24 24"
+              class="h-full w-full fill-current text-current"
+            >
+              <path :d="siMisskey.path" />
+            </svg>
+          </SocialShareButton>
+
+          <Transition
+            enter-active-class="transition duration-200 ease-out"
+            enter-from-class="translate-y-1 opacity-0"
+            enter-to-class="translate-y-0 opacity-100"
+            leave-active-class="transition duration-150 ease-in"
+            leave-from-class="translate-y-0 opacity-100"
+            leave-to-class="translate-y-1 opacity-0"
+          >
+            <SocialSharePanel
+              :class="[
+                'absolute bottom-12 left-1/2 -translate-x-1/2',
+                'grid w-52 gap-1 rounded-lg p-3',
+                'shadow-lg dark:shadow-slate-950/70',
+                'bg-white dark:bg-slate-800',
+              ]"
+            >
+              <label class="select-none text-xs" for="share-misskey">
+                サーバのドメイン：
+              </label>
+              <div class="flex items-center gap-2">
+                <SocialShareInput
+                  id="share-misskey"
+                  placeholder="example.com"
+                  :class="[
+                    'shrink grow basis-full',
+                    'w-full rounded px-2 py-1 text-sm outline-none',
+                    'bg-slate-100 dark:bg-slate-700',
+                  ]"
+                />
+                <SocialShareSubmit
+                  :class="[
+                    'shrink-0 rounded border border-slate-300 px-2 py-1 text-sm',
+                  ]"
+                >
+                  投稿
+                </SocialShareSubmit>
+              </div>
+            </SocialSharePanel>
+          </Transition>
+        </SocialShare>
+      </div>
+    </div>
+
+    <div
+      v-if="page?._dir === 'blog'"
+      class="flex flex-wrap items-start justify-between gap-x-12 gap-y-6"
+    >
+      <NuxtLink
+        v-if="prev && prev?._dir === 'blog' && prev?._path"
+        :to="prev?._path?.endsWith('/') ? prev?._path : `${prev?._path}/`"
+        :class="[
+          'flex grow basis-1/2 items-center justify-start gap-4',
+          'relative mr-auto max-w-xs pb-2',
+          'after:absolute after:bottom-0 after:left-0',
+          'after:bg-primary after:block after:h-0.5 after:w-full',
+          'after:scale-x-0 after:scale-y-100',
+          'after:origin-right after:transition-transform',
+          'hover:after:origin-left hover:after:scale-x-100',
+        ]"
+      >
+        <ChevronLeftIcon class="h-5 w-5" aria-hidden="true" />
+        <article class="flex flex-col items-start gap-1">
+          <span class="text-xs text-slate-600 dark:text-slate-400">
+            前の投稿
+          </span>
+          <h2 class="text-sm font-bold">
+            {{ prev?.title }}
+          </h2>
+        </article>
+      </NuxtLink>
+      <NuxtLink
+        v-if="next && next?._dir === 'blog' && next?._path"
+        :to="next?._path?.endsWith('/') ? next?._path : `${next?._path}/`"
+        :class="[
+          'flex grow basis-1/2 items-center justify-end gap-4',
+          'relative ml-auto max-w-xs pb-2',
+          'after:absolute after:bottom-0 after:left-0',
+          'after:bg-primary after:block after:h-0.5 after:w-full',
+          'after:scale-x-0 after:scale-y-100',
+          'after:origin-right after:transition-transform',
+          'hover:after:origin-left hover:after:scale-x-100',
+        ]"
+      >
+        <article class="flex flex-col items-end gap-1">
+          <span class="text-xs text-slate-600 dark:text-slate-400">
+            次の投稿
+          </span>
+          <h2 class="text-right text-sm font-bold">
+            {{ next?.title }}
+          </h2>
+        </article>
+        <ChevronRightIcon class="h-5 w-5" aria-hidden="true" />
+      </NuxtLink>
+    </div>
+  </footer>
+</template>
